Move the sending of e-mail to the new application context
