+++ src/external/cli/cliApplicationContext.go
@@ -1,11 +1,10 @@
 package cli
 
 import (
-	"github.com/halprin/email-conceal/forwarder/gateways"
-	"github.com/halprin/email-conceal/forwarder/usecases"
 	"github.com/halprin/email-conceal/src/context"
 	forwardEmailController "github.com/halprin/email-conceal/src/controllers/forwardEmail"
 	"github.com/halprin/email-conceal/src/external/lib"
+	"github.com/halprin/email-conceal/src/gateways/awsSesSendEmail"
 	"github.com/halprin/email-conceal/src/gateways/dynamodb"
 	"github.com/halprin/email-conceal/src/gateways/localFileReader"
 	"github.com/halprin/email-conceal/src/gateways/osEnvironmentVariable"
@@ -59,6 +58,10 @@ func init() {
 	//gateways
 	applicationContext.Bind(func() forwardEmailUsecase.ReadEmailGateway {
 		return localFileReader.LocalFileReader{}
+	})
+
+	applicationContext.Bind(func() forwardEmailUsecase.SendEmailGateway {
+		return awsSesSendEmail.AwsSesSendEmailGateway{}
 	})
 
 	applicationContext.Bind(func() context.EnvironmentGateway {
