+++ src/gateways/awsSesSendEmail/awsSession.go
@@ -1,0 +1,5 @@
+package awsSesSendEmail
+
+import "github.com/aws/aws-sdk-go/aws/session"
+
+var awsSession, sessionErr = session.NewSession()
