+++ src/usecases/forwardEmail/gatewayInterfaces.go
@@ -3,3 +3,7 @@ package forwardEmail
 type ReadEmailGateway interface {
 	ReadEmail(uri string) ([]byte, error)
 }
+
+type SendEmailGateway interface {
+	SendEmail(email []byte, recipients []string) error
+}
